+++ Runtime/Extensions/UnityWebRequestExtensions.cs
@@ -1,0 +1,12 @@
+using UnityEngine.Networking;
+
+namespace EZ.Network.Extensions
+{
+    public static class UnityWebRequestExtensions
+    {
+        public static UnityWebRequestAwaiter GetAwaiter(this UnityWebRequestAsyncOperation asyncOperation)
+        {
+            return new UnityWebRequestAwaiter(asyncOperation);
+        }
+    }
+}