+++ Runtime/ProtocolFlags.cs
@@ -1,0 +1,10 @@
+namespace EZ.Network
+{
+    [System.Flags]
+    public enum Protocol
+    {
+        Http = 1 << 1,
+        Tcp = 1 << 2,
+        Everything = ~0
+    }
+}