+++ Runtime/UnityWebRequestAwaiter.cs
@@ -1,0 +1,33 @@
+using System;
+using System.Runtime.CompilerServices;
+using UnityEngine;
+using UnityEngine.Networking;
+
+namespace EZ.Network
+{
+    public class UnityWebRequestAwaiter : INotifyCompletion
+    {
+        private UnityWebRequestAsyncOperation _asyncOperation;
+        private Action _continuation;
+
+        public UnityWebRequestAwaiter(UnityWebRequestAsyncOperation asyncOperation)
+        {
+            _asyncOperation = asyncOperation;
+            asyncOperation.completed += OnRequestCompleted;
+        }
+
+        public bool IsCompleted { get { return _asyncOperation.isDone; } }
+
+        public void GetResult() { }
+
+        public void OnCompleted(Action continuation)
+        {
+            _continuation = continuation;
+        }
+
+        private void OnRequestCompleted(AsyncOperation asyncOperation)
+        {
+            _continuation();
+        }
+    }
+}